requirements.txt
pyusb==1.0.2
click==7.0
sounddevice
numpy
openai-whisper
scipy